+++ FindPanel/OgreTextFindComponents/NSOutlineView/OgreOutlineViewAdapter.m
@@ -163,7 +163,7 @@
 #ifdef DEBUG_OGRE_FIND_PANEL
 	NSLog(@"  -numberOfChildrenInSelection: of %@", [self className]);
 #endif
-    int count = [_outlineView numberOfSelectedColumns];
+    NSInteger count = [_outlineView numberOfSelectedColumns];
     if (inSelection && (count > 0)) return count;
     
     return [_outlineView numberOfColumns];
@@ -220,7 +220,7 @@
 #ifdef DEBUG_OGRE_FIND_PANEL
 	NSLog(@"  -componentEnumeratorInSelection: of %@", [self className]);
 #endif
-    int count = [_outlineView numberOfSelectedColumns];
+    NSInteger count = [_outlineView numberOfSelectedColumns];
     
     OgreTextFindComponentEnumerator *enumerator;
     if ([self isReversed]) {
