+++ FindPanel/OgreTextFindThreads/OgreReplaceAllThread.h
@@ -21,7 +21,7 @@
 {
     NSArray					*matchArray;
     OGReplaceExpression		*repex;
-    unsigned				aNumberOfReplaces, aNumberOfMatches;
+    NSUInteger				aNumberOfReplaces, aNumberOfMatches;
     NSString				*progressMessage, *progressMessagePlural, *remainingTimeMesssage;
 	NSObject<OGStringProtocol>				*replacedString;
 }
